Upgrade to stream_transform v0.0.6
Fixes https://github.com/dart-lang/stream_transform/issues/18

This version has a fix for `debounce` to forward events to all
listeners.

Also updates the test to start listening for the first value on the
stream before typing through the page object since typing also waits for
the Timer to fire.